+++ packages/keyrecall_practice/test/failure_atomicity_test.dart
@@ -64,6 +64,14 @@ class FlakyPracticeStore implements PracticeStore {
       inner.saveCheckpoint(checkpoint);
 
   @override
+  Future<List<CoordinationSample>> loadCoordinationSamples(String profileId) =>
+      inner.loadCoordinationSamples(profileId);
+
+  @override
+  Future<void> appendCoordinationSample(CoordinationSample sample) =>
+      inner.appendCoordinationSample(sample);
+
+  @override
   Future<PracticePlan?> loadPracticePlan(String profileId) =>
       inner.loadPracticePlan(profileId);
 
@@ -121,6 +129,14 @@ class MisfilingPracticeStore implements PracticeStore {
   @override
   Future<void> saveCheckpoint(LearnerStateCheckpoint checkpoint) =>
       inner.saveCheckpoint(checkpoint);
+
+  @override
+  Future<List<CoordinationSample>> loadCoordinationSamples(String profileId) =>
+      inner.loadCoordinationSamples(profileId);
+
+  @override
+  Future<void> appendCoordinationSample(CoordinationSample sample) =>
+      inner.appendCoordinationSample(sample);
 
   @override
   Future<PracticePlan?> loadPracticePlan(String profileId) =>